// src/api/generated/models/municipalityResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Gaula Digital — API
 * Sistema de información operacional del GAULA, Ejército Nacional de Colombia.
 * OpenAPI spec version: v1
 */

export interface MunicipalityResponse {
  code?: string;
  name?: string;
  departmentName?: string;
  territorialUnitName?: string;
}